feat!: make `signal.get` reactive and add `signal.getOnce`
@affects atoms, react, stores
