+++ packages/atoms/src/classes/SelectorInstance.ts
@@ -13,15 +13,15 @@ import {
   startBuffer,
 } from '../utils/evaluationContext'
 import { prefix } from '../utils/general'
-import { pluginActions } from '../utils/plugin-actions'
-import { Ecosystem } from './Ecosystem'
 import {
   destroyNodeFinish,
   destroyNodeStart,
-  GraphNode,
   scheduleDependents,
   setNodeStatus,
-} from './GraphNode'
+} from '../utils/graph'
+import { pluginActions } from '../utils/plugin-actions'
+import { Ecosystem } from './Ecosystem'
+import { GraphNode } from './GraphNode'
 
 const defaultResultsComparator = (a: any, b: any) => a === b
 
@@ -129,11 +129,6 @@ export class SelectorInstance<
 > extends GraphNode<G & { Events: any }> {
   public static $$typeof = Symbol.for(`${prefix}/SelectorInstance`)
 
-  /**
-   * `v`alue - the current cached selector result
-   */
-  public v?: G['State']
-
   constructor(
     /**
      * @see GraphNode.e
@@ -156,6 +151,7 @@ export class SelectorInstance<
     public p: G['Params']
   ) {
     super()
+    runSelector(this, true)
   }
 
   /**
@@ -170,17 +166,10 @@ export class SelectorInstance<
   }
 
   /**
-   * @see GraphNode.get
-   */
-  public get() {
-    return this.v as G['State']
-  }
-
-  /**
    * @see GraphNode.d
    */
   public d(options?: DehydrationFilter) {
-    if (this.f(options)) return this.get()
+    if (this.f(options)) return this.v
   }
 
   /**
